Module syntax_pos::span_encoding

Structs

Span
A compressed span. Contains either fields of SpanData inline if they are small, or index into span interner. The primary goal of Span is to be as small as possible and fit into other structures (that's why it uses packed as well). Decoding speed is the second priority. See SpanData for the info on span fields in decoded representation.

SpanInterner

Constants

BASE_INDEX

CTXT_INDEX

DUMMY_SP
Dummy span, both position and length are zero, syntax context is zero as well. This span is kept inline and encoded with format 0.

INLINE_OFFSETS

INLINE_SIZES

INTERNED_INDEX_OFFSET

INTERNED_INDEX_SIZE

LEN_INDEX

TAG_INLINE

TAG_INTERNED

TAG_MASK

Functions

decode

encode

with_span_interner