Struct syntax_pos::MultiSpan

pub struct MultiSpan { /* fields omitted */ }

This is an experimental API. (rustc_private)

A collection of spans. Spans have two orthogonal attributes:

they can be primary spans. In this case they are the locus of the error, and would be rendered with ^^^.

they can have a label. In this case, the label is written next to the mark in the snippet when we render.

Methods

impl MultiSpan

fn new() -> MultiSpan

This is an experimental API. (rustc_private)

fn from_span(primary_span: Span) -> MultiSpan

This is an experimental API. (rustc_private)

fn from_spans(vec: Vec<Span>) -> MultiSpan

This is an experimental API. (rustc_private)

fn push_span_label(&mut self, span: Span, label: String)

This is an experimental API. (rustc_private)

fn primary_span(&self) -> Option<Span>

This is an experimental API. (rustc_private)

Selects the first primary span (if any)

fn primary_spans(&self) -> &[Span]

This is an experimental API. (rustc_private)

Returns all primary spans.

fn replace(&mut self, before: Span, after: Span) -> bool

This is an experimental API. (rustc_private)

Replaces all occurances of one Span with another. Used to move Spans in areas that don't display well (like std macros). Returns true if replacements occurred.

fn span_labels(&self) -> Vec<SpanLabel>

This is an experimental API. (rustc_private)

Returns the strings to highlight. We always ensure that there is an entry for each of the primary spans -- for each primary span P, if there is at least one label with span P, we return those labels (marked as primary). But otherwise we return SpanLabel instances with empty labels.