Crate syntax_pos

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The source positions and related helper functions

Note

This API is completely unstable and subject to change.

Reexports

pub use hygiene::SyntaxContext;

pub use hygiene::ExpnInfo;

pub use hygiene::ExpnFormat;

pub use hygiene::NameAndSpan;

pub use hygiene::CompilerDesugaringKind;

Modules

hygiene 	 [
Experimental
]
Machinery for hygienic macros, inspired by the MTWT[1] paper.

symbol 	 [
Experimental
]
An "interner" is a data structure that associates values with usize tags and allows bidirectional lookup; i.e. given a value, one can easily find the type, and vice versa.

Structs

BytePos 	 [
Experimental
]
A byte offset. Keep this small (currently 32-bits), as AST contains a lot of them.

CharPos 	 [
Experimental
]
A character offset. Because of multibyte utf8 characters, a byte offset is not equivalent to a character offset. The CodeMap will convert BytePos values to CharPos values as necessary.

DistinctSources 	 [
Experimental
]

FileLines 	 [
Experimental
]

FileMap 	 [
Experimental
]
A single source in the CodeMap.

FileMapAndBytePos 	 [
Experimental
]

FileMapAndLine 	 [
Experimental
]

LineInfo 	 [
Experimental
]

Loc 	 [
Experimental
]
A source code location used for error reporting

LocWithOpt 	 [
Experimental
]
A source code location used as the result of lookup_char_pos_adj

MacroBacktrace 	 [
Experimental
]

MalformedCodemapPositions 	 [
Experimental
]

MultiByteChar 	 [
Experimental
]
Identifies an offset of a multi-byte character in a FileMap

MultiSpan 	 [
Experimental
]
A collection of spans. Spans have two orthogonal attributes:

Span 	 [
Experimental
]
A compressed span. Contains either fields of SpanData inline if they are small, or index into span interner. The primary goal of Span is to be as small as possible and fit into other structures (that's why it uses packed as well). Decoding speed is the second priority. See SpanData for the info on span fields in decoded representation.

SpanData 	 [
Experimental
]
Spans represent a region of code, used for error reporting. Positions in spans are absolute positions from the beginning of the codemap, not positions relative to FileMaps. Methods on the CodeMap can be used to relate spans back to the original source. You must be careful if the span crosses more than one file - you will not be able to use many of the functions on spans in codemap and you cannot assume that the length of the span = hi - lo; there may be space in the BytePos range between files.

SpanLabel 	 [
Experimental
]

Enums

ExternalSource 	 [
Experimental
]
The state of the lazy external source loading mechanism of a FileMap.

SpanLinesError 	 [
Experimental
]

SpanSnippetError 	 [
Experimental
]

Constants

DUMMY_SP 	 [
Experimental
]
Dummy span, both position and length are zero, syntax context is zero as well. This span is kept inline and encoded with format 0.

NO_EXPANSION 	 [
Experimental
]

SPAN_DEBUG 	 [
Experimental
]

Traits

Pos 	 [
Experimental
]

Type Definitions

FileLinesResult 	 [
Experimental
]

FileName 	 [
Experimental
]